requirements_api.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
requests
pillow
python-dotenv
